requirements.txt
aio-btdht @ git+https://github.com/bashkirtsevich-llc/aiobtdht@f99d881e9b992da13519721e251912c32a566cd7
aio-krpc-server @ git+https://github.com/bashkirtsevich-llc/aiokrpc@fe0b18a98d4dc19034f8152afbae78fb600830e2
aio-udp-server==0.0.6
beautifulsoup4==4.12.2
bencode.py==4.0.0
bs4==0.0.1
Cerberus==1.3.1
certifi==2023.11.17
charset-normalizer==3.3.2
humanize==4.9.0
idna==3.6
Jinja2==3.1.2
MarkupSafe==2.1.3
peewee==3.17.0
py3-bencode==0.0.3
redis==5.0.1
requests==2.31.0
soupsieve==2.5
toml==0.10.2
tracker-scraper==1.1.0
urllib3==2.1.0
